requirements.txt
streamlit
google-generativeai>=0.2.0  # Ensure compatibility with newer models like gemini-1.5-flash
Pillow
python-dotenv
